requirements.txt
adrf==0.1.7
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1
wheel==0.43.0
